Cleanup memtier library
- use only memtier kind object on memtier layer
