+++ tiering/memtier.c
@@ -10,9 +10,6 @@
 #include <pthread.h>
 #include <string.h>
 
-// TODO - remove this after logging cleanup
-#include <memkind/internal/memkind_private.h>
-
 #define MEMTIER_EXPORT __attribute__((visibility("default")))
 #define MEMTIER_INIT   __attribute__((constructor))
 #define MEMTIER_FINI   __attribute__((destructor))
@@ -22,9 +19,7 @@
 
 static int destructed;
 
-// TODO create structure to keep kind + multiple tiers
 static struct memtier_kind *current_kind;
-static struct memtier_tier *current_tier;
 
 MEMTIER_EXPORT void *malloc(size_t size)
 {
@@ -100,99 +95,6 @@ MEMTIER_EXPORT void free(void *ptr)
     }
 }
 
-static memkind_t get_kind(const char *str, const char *pmem_path,
-                          size_t pmem_size)
-{
-    memkind_t kind = NULL;
-    if (strcmp(str, "DRAM") == 0) {
-        kind = MEMKIND_DEFAULT;
-    } else if (strcmp(str, "FS_DAX") == 0) {
-        // TODO handle FS_DAX here
-    }
-
-    log_debug("kind_name: %s", kind->name);
-    log_debug("pmem_path: %s", pmem_path);
-    log_debug("pmem_size: %zu", pmem_size);
-
-    return kind;
-}
-
-// TODO - move this code during refactoring after implementing support for
-// multiple kinds
-static const char *policy_to_str(memtier_policy_t policy)
-{
-    if (policy < MEMTIER_POLICY_CIRCULAR ||
-        policy >= MEMTIER_POLICY_MAX_VALUE) {
-        log_err("Unknown policy: %d", policy);
-        return NULL;
-    }
-
-    const char *policies[] = {"POLICY_CIRCULAR"};
-
-    return policies[policy];
-}
-
-// TODO move this logic to ctl
-static int create_tiered_kind_from_env(char *env_var_string)
-{
-    char *kind_name = NULL;
-    char *pmem_path = NULL;
-    size_t pmem_size;
-    unsigned ratio_value = 0;
-    memtier_policy_t policy = MEMTIER_POLICY_MAX_VALUE;
-
-    int ret = ctl_load_config(env_var_string, &kind_name, &pmem_path,
-                              &pmem_size, &ratio_value, &policy);
-    if (ret != 0) {
-        return -1;
-    }
-
-    memkind_t kind = get_kind(kind_name, pmem_path, pmem_size);
-    if (kind == NULL) {
-        return -1;
-    }
-
-    log_debug("ratio_value: %u", ratio_value);
-    log_debug("policy: %s", policy_to_str(policy));
-
-    current_tier = memtier_tier_new(kind);
-    if (current_tier == NULL) {
-        return -1;
-    }
-
-    struct memtier_builder *builder = memtier_builder_new();
-    if (!builder) {
-        ret = -1;
-        goto tier_delete;
-    }
-
-    ret = memtier_builder_add_tier(builder, current_tier, ratio_value);
-    if (ret != 0) {
-        goto builder_delete;
-    }
-
-    ret = memtier_builder_set_policy(builder, policy);
-    if (ret != 0) {
-        goto builder_delete;
-    }
-
-    ret = memtier_builder_construct_kind(builder, &current_kind);
-    if (ret != 0) {
-        goto builder_delete;
-    }
-    memtier_builder_delete(builder);
-
-    return ret;
-
-builder_delete:
-    memtier_builder_delete(builder);
-
-tier_delete:
-    memtier_tier_delete(current_tier);
-
-    return ret;
-}
-
 static pthread_once_t init_once = PTHREAD_ONCE_INIT;
 
 static MEMTIER_INIT void memtier_init(void)
@@ -202,8 +104,8 @@ static MEMTIER_INIT void memtier_init(void)
 
     char *env_var = utils_get_env("MEMKIND_MEM_TIERING_CONFIG");
     if (env_var) {
-        int ret = create_tiered_kind_from_env(env_var);
-        if (!ret) {
+        current_kind = ctl_create_tier_kind_from_env(env_var);
+        if (current_kind) {
             return;
         }
         log_err("Error with parsing MEMKIND_MEM_TIERING_CONFIG");
@@ -217,12 +119,8 @@ static MEMTIER_FINI void memtier_fini(void)
 {
     log_info("Unloading memkind memtier lib!");
 
-    if (current_kind) {
-        // TODO currently we handle one tier
-        memtier_tier_delete(current_tier);
-        memtier_delete_kind(current_kind);
-        current_kind = NULL;
-    }
+    ctl_destroy_kind(current_kind);
+    current_kind = NULL;
 
     destructed = 1;
 }
